chore: add pre-commit config with rust fmt + clippy hooks
Uses pre-commit.com with doublify/pre-commit-rust for:
- cargo fmt --all
- cargo clippy --all-targets --all-features -- -D warnings
- check-merge-conflict, trailing-whitespace, end-of-file-fixer
- check-added-large-files (500KB)

Also fixes two clippy issues caught by --all-targets:
- unnecessary i64 cast in test
- iter().copied().collect() → to_vec()

+++ datafusion/bio-function-vep/src/cache_builder.rs
@@ -3488,7 +3488,7 @@ mod tests {
             pending.push(PositionAccumulator::new(
                 1,
                 "1".to_string(),
-                i as i64 * 10,
+                i * 10,
                 0,
                 &batch,
             ));
